move updates/uploader into commits pkg, add many_many gorm relationship for commits+updates
